refactor: rename graph components to kebab-case
