+++ ui/components/graphs/horizontal-bar-chart.tsx
