+++ ui/components/graphs/line-chart.tsx
@@ -14,8 +14,8 @@ import {
   YAxis,
 } from "recharts";
 
-import { AlertPill } from "./shared/AlertPill";
-import { ChartLegend } from "./shared/ChartLegend";
+import { AlertPill } from "./shared/alert-pill";
+import { ChartLegend } from "./shared/chart-legend";
 import { CHART_COLORS } from "./shared/constants";
 import { LineConfig, LineDataPoint } from "./types";
