+++ ui/components/graphs/shared/chart-legend.tsx
